[www]: (refactor): Update site header, Remove UI from packages, Rethink things
