+++ apps/www/src/app/(app)/layout.tsx
@@ -1,0 +1,14 @@
+import { SiteHeader } from "@/app/(app)/_components/side-header";
+
+interface AppLayoutProps {
+    children: React.ReactNode;
+}
+
+export default function AppLayout({ children }: AppLayoutProps) {
+    return (
+        <>
+            <SiteHeader />
+            <main className="flex-1">{children}</main>
+        </>
+    );
+}
