+++ apps/www/src/components/theme-toggle.tsx
@@ -1,41 +1,37 @@
 "use client";
 
 import * as React from "react";
-import { MoonIcon, SunIcon } from "lucide-react";
 import { useTheme } from "next-themes";
 
-import { Button } from "@rapidlaunch/ui/button";
 import {
-  DropdownMenu,
-  DropdownMenuCheckboxItem,
-  DropdownMenuContent,
-  DropdownMenuTrigger,
-} from "@rapidlaunch/ui/dropdown";
+    DropdownMenu,
+    DropdownMenuCheckboxItem,
+    DropdownMenuContent,
+    DropdownMenuTrigger,
+} from "@/components/ui/dropdown-menu";
 
-export function ThemeToggle() {
-  const { setTheme, theme, themes } = useTheme();
+interface ThemeToggleProps {
+    button: React.ReactNode;
+}
 
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button variant="outline" size="icon">
-          <SunIcon className="h-[1.2rem] w-[1.2rem] rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0" />
-          <MoonIcon className="absolute h-[1.2rem] w-[1.2rem] rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-100" />
-          <span className="sr-only">Toggle theme</span>
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="end">
-        {themes.map((t) => (
-          <DropdownMenuCheckboxItem
-            key={t}
-            checked={theme === t}
-            onClick={() => setTheme(t)}
-            className="text-sm capitalize"
-          >
-            {t}
-          </DropdownMenuCheckboxItem>
-        ))}
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
+export function ThemeToggle({ button }: ThemeToggleProps) {
+    const { setTheme, theme, themes } = useTheme();
+
+    return (
+        <DropdownMenu>
+            <DropdownMenuTrigger asChild>{button}</DropdownMenuTrigger>
+            <DropdownMenuContent align="end">
+                {themes.map((t) => (
+                    <DropdownMenuCheckboxItem
+                        key={t}
+                        checked={theme === t}
+                        onClick={() => setTheme(t)}
+                        className="text-sm capitalize"
+                    >
+                        {t}
+                    </DropdownMenuCheckboxItem>
+                ))}
+            </DropdownMenuContent>
+        </DropdownMenu>
+    );
 }
