+++ apps/www/src/components/ui/form.tsx
@@ -1,29 +1,23 @@
-import * as LabelPrimitive from "@radix-ui/react-label";
+import * as React from "react";
+import type * as LabelPrimitive from "@radix-ui/react-label";
 import { Slot } from "@radix-ui/react-slot";
-import {
-    Controller,
-    ControllerProps,
-    FieldPath,
-    FieldValues,
-    FormProvider,
-    useFormContext,
-} from "react-hook-form";
+import { Controller, FormProvider, useFormContext } from "react-hook-form";
+import type { ControllerProps, FieldPath, FieldValues } from "react-hook-form";
 
-import { cn } from ".";
-import { Label } from "./label";
-import { createContext, forwardRef, useContext, useId } from "react";
+import { cn } from "@/lib/utils";
+import { Label } from "@/components/ui/label";
 
 const Form = FormProvider;
 
-type FormFieldContextValue<
+interface FormFieldContextValue<
     TFieldValues extends FieldValues = FieldValues,
     TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>,
-> = {
+> {
     name: TName;
-};
+}
 
-const FormFieldContext = createContext<FormFieldContextValue>(
-    {} as FormFieldContextValue
+const FormFieldContext = React.createContext<FormFieldContextValue>(
+    {} as FormFieldContextValue,
 );
 
 const FormField = <
@@ -40,8 +34,8 @@ const FormField = <
 };
 
 const useFormField = () => {
-    const fieldContext = useContext(FormFieldContext);
-    const itemContext = useContext(FormItemContext);
+    const fieldContext = React.useContext(FormFieldContext);
+    const itemContext = React.useContext(FormItemContext);
     const { getFieldState, formState } = useFormContext();
 
     const fieldState = getFieldState(fieldContext.name, formState);
@@ -62,19 +56,19 @@ const useFormField = () => {
     };
 };
 
-type FormItemContextValue = {
+interface FormItemContextValue {
     id: string;
-};
+}
 
-const FormItemContext = createContext<FormItemContextValue>(
-    {} as FormItemContextValue
+const FormItemContext = React.createContext<FormItemContextValue>(
+    {} as FormItemContextValue,
 );
 
-const FormItem = forwardRef<
+const FormItem = React.forwardRef<
     HTMLDivElement,
     React.HTMLAttributes<HTMLDivElement>
 >(({ className, ...props }, ref) => {
-    const id = useId();
+    const id = React.useId();
 
     return (
         <FormItemContext.Provider value={{ id }}>
@@ -84,7 +78,7 @@ const FormItem = forwardRef<
 });
 FormItem.displayName = "FormItem";
 
-const FormLabel = forwardRef<
+const FormLabel = React.forwardRef<
     React.ElementRef<typeof LabelPrimitive.Root>,
     React.ComponentPropsWithoutRef<typeof LabelPrimitive.Root>
 >(({ className, ...props }, ref) => {
@@ -101,7 +95,7 @@ const FormLabel = forwardRef<
 });
 FormLabel.displayName = "FormLabel";
 
-const FormControl = forwardRef<
+const FormControl = React.forwardRef<
     React.ElementRef<typeof Slot>,
     React.ComponentPropsWithoutRef<typeof Slot>
 >(({ ...props }, ref) => {
@@ -124,7 +118,7 @@ const FormControl = forwardRef<
 });
 FormControl.displayName = "FormControl";
 
-const FormDescription = forwardRef<
+const FormDescription = React.forwardRef<
     HTMLParagraphElement,
     React.HTMLAttributes<HTMLParagraphElement>
 >(({ className, ...props }, ref) => {
@@ -134,14 +128,14 @@ const FormDescription = forwardRef<
         <p
             ref={ref}
             id={formDescriptionId}
-            className={cn("text-[0.8rem] text-muted-foreground", className)}
+            className={cn("text-sm text-muted-foreground", className)}
             {...props}
         />
     );
 });
 FormDescription.displayName = "FormDescription";
 
-const FormMessage = forwardRef<
+const FormMessage = React.forwardRef<
     HTMLParagraphElement,
     React.HTMLAttributes<HTMLParagraphElement>
 >(({ className, children, ...props }, ref) => {
@@ -156,10 +150,7 @@ const FormMessage = forwardRef<
         <p
             ref={ref}
             id={formMessageId}
-            className={cn(
-                "text-[0.8rem] font-medium text-destructive",
-                className
-            )}
+            className={cn("text-sm font-medium text-destructive", className)}
             {...props}
         >
             {body}
