+++ apps/www/src/config/nav.ts
@@ -1,0 +1,24 @@
+export const navConfig = {
+    items: [
+        {
+            label: "Starterkits",
+            href: "/starterkits",
+        },
+        {
+            label: "Components",
+            href: "/components",
+        },
+        {
+            label: "Blog",
+            href: "/blog",
+        },
+        {
+            label: "Docs",
+            href: "/docs",
+        },
+        {
+            label: "GitHub",
+            href: "/gh",
+        },
+    ],
+};
