+++ apps/www/src/config/urls.ts
@@ -1,5 +1,9 @@
 export const siteUrls = {
     home: "/",
+    starterkits: "/starterkits",
+    components: "/components",
+    blog: "/blog",
+    docs: "/docs",
     twitter: "https://twitter.com/AliFarooqDev",
-    github: "https://github.com/alifarooq9/rapidlaunch",
+    github: "/gh",
 } as const;
