--- apps/www/src/hooks/use-mouse-move.ts
@@ -1,25 +1,0 @@
-"use client";
-
-import React from "react";
-
-export default function useMouseMove() {
-  React.useEffect(() => {
-    function mouseMoveEvent(e: MouseEvent) {
-      const scale = window.visualViewport?.scale;
-      if (scale === 1) {
-        const body = document.body;
-
-        const targetX = e.clientX;
-        const targetY = e.clientY;
-
-        body.style.setProperty("--x", `${targetX}px`);
-        body.style.setProperty("--y", `${targetY}px`);
-      }
-    }
-
-    document.addEventListener("mousemove", mouseMoveEvent);
-    return () => {
-      document.removeEventListener("mousemove", mouseMoveEvent);
-    };
-  }, []);
-}
