+++ apps/www/src/lib/fonts.ts
@@ -1,0 +1,5 @@
+import { Inter } from "next/font/google";
+
+export const fontSans = Inter({
+    subsets: ["latin"],
+});
