--- packages/ui/package.json
@@ -1,45 +1,0 @@
-{
-    "name": "@rapidlaunch/ui",
-    "version": "0.0.0",
-    "private": true,
-    "exports": {
-        ".": "./src/index.ts",
-        "./button": "./src/button.tsx",
-        "./dropdown": "./src/dropdown.tsx",
-        "./badge": "./src/badge.tsx",
-        "./input": "./src/input.tsx",
-        "./label": "./src/label.tsx",
-        "./form": "./src/form.tsx",
-        "./sonner": "./src/sonner.tsx"
-    },
-    "scripts": {
-        "lint": "eslint .",
-        "generate:component": "turbo gen react-component"
-    },
-    "devDependencies": {
-        "@rapidlaunch/eslint-config": "workspace:^",
-        "@rapidlaunch/tailwind-config": "workspace:*",
-        "@rapidlaunch/typescript-config": "workspace:*",
-        "@turbo/gen": "^1.12.4",
-        "@types/eslint": "^8.56.5",
-        "@types/node": "^20.11.24",
-        "@types/react": "^18.2.63",
-        "@types/react-dom": "^18.2.19",
-        "eslint": "^8.57.0",
-        "react": "^18.2.0",
-        "tailwindcss": "^3.4.1",
-        "typescript": "^5.3.3"
-    },
-    "dependencies": {
-        "@radix-ui/react-dropdown-menu": "^2.0.6",
-        "@radix-ui/react-icons": "^1.3.0",
-        "@radix-ui/react-label": "^2.0.2",
-        "@radix-ui/react-slot": "^1.0.2",
-        "class-variance-authority": "^0.7.0",
-        "clsx": "^2.1.0",
-        "next-themes": "^0.2.1",
-        "react-hook-form": "^7.50.1",
-        "sonner": "^1.4.2",
-        "tailwind-merge": "^2.2.1"
-    }
-}
