--- packages/ui/src/badge.tsx
@@ -1,48 +1,0 @@
-import * as React from "react";
-import { cva, type VariantProps } from "class-variance-authority";
-
-import { cn } from ".";
-
-const badgeVariants = cva(
-    "inline-flex items-center rounded-md border font-semibold transition-colors focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2",
-    {
-        variants: {
-            variant: {
-                default:
-                    "border-transparent bg-primary text-primary-foreground shadow hover:bg-primary/80",
-                secondary:
-                    "border-transparent bg-secondary text-secondary-foreground hover:bg-secondary/80",
-                destructive:
-                    "border-transparent bg-destructive text-destructive-foreground shadow hover:bg-destructive/80",
-                outline: "text-foreground",
-                background: "bg-background text-background-foreground",
-                success: "border-transparent bg-green-500 text-white",
-                info: "border-transparent bg-yellow-200 text-black",
-            },
-            size: {
-                sm: "text-xs px-2 py-0.5",
-                md: "text-sm px-2.5 py-1",
-                lg: "text-base px-4 py-2",
-            },
-        },
-        defaultVariants: {
-            variant: "default",
-            size: "sm",
-        },
-    }
-);
-
-export interface BadgeProps
-    extends React.HTMLAttributes<HTMLDivElement>,
-        VariantProps<typeof badgeVariants> {}
-
-function Badge({ className, variant, size, ...props }: BadgeProps) {
-    return (
-        <div
-            className={cn(badgeVariants({ variant, size }), className)}
-            {...props}
-        />
-    );
-}
-
-export { Badge, badgeVariants };
