--- packages/ui/tailwind.config.ts
@@ -1,12 +1,0 @@
-/*
- * This file is not used for any compilation purpose, it is only used
- * for Tailwind Intellisense & Autocompletion in the source files
- */
-import type { Config } from "tailwindcss";
-
-import baseConfig from "@rapidlaunch/tailwind-config/base";
-
-export default {
-    content: ["./src/**/*.tsx"],
-    presets: [baseConfig],
-} satisfies Config;
