--- packages/ui/tsconfig.json
@@ -1,8 +1,0 @@
-{
-    "extends": "@rapidlaunch/typescript-config/react-library.json",
-    "compilerOptions": {
-        "tsBuildInfoFile": "node_modules/.cache/tsbuildinfo.json"
-    },
-    "include": ["src", "."],
-    "exclude": ["node_modules", "dist"]
-}
