--- packages/ui/tsconfig.lint.json
@@ -1,8 +1,0 @@
-{
-    "extends": "@rapidlaunch/typescript-config/react-library.json",
-    "compilerOptions": {
-        "tsBuildInfoFile": "node_modules/.cache/tsbuildinfo.json"
-    },
-    "include": ["src", "turbo", "."],
-    "exclude": ["node_modules", "dist"]
-}
